Complete rewrite and restructure, adding config, split into multiple py
