+++ .gitignore
@@ -1,4 +1,4 @@
 *.dat
 *.d1c
-/*
+*.pyc
 *.txt
