+++ README.md
@@ -1,2 +1,2 @@
 # PT-Generator
-Generates files that are able to be read into an end user's database.
+Generates files that are inputs for back end accounting software.
